Move InvokeAsync to Reference
